🔧 refactor: Simplify prediction reading and update relationships in Paragraph and Prediction models
